[not verified] Apply widget filtering to customizer preview
The customizer screen was not applying saved visibility rules for
widgets in the preview on the right hand side. This change addresses
that by distinguishing between the customizer and the preview using the
`customize_changeset_uuid` GET var.

Note that when widget visibility rules are first added they will not be
respected until the preview is reloaded, e.g. by navigating inside the
preview or by refreshing page.

Addresses #21446